backend/requirements.txt
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
setuptools==80.9.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
